# Reuse float histogram objects