This commit builds upon https://github.com/prometheus/prometheus/pull/13215, and reduces memory for querying native histogram by reusing existing HPoint instances.

Benchmarks show ~23% improvement in performance and ~45% reduction in memory usage. Queries without `rate` are not affected yet. To make improvements there we will need to change the iterator interface and I will do that in a follow up PR.

```
benchstat main.out new.out                                                                    
name                         old time/op    new time/op    delta
NativeHistograms/sum-8          521ms ± 2%     522ms ± 3%     ~     (p=1.000 n=5+5)
NativeHistograms/sum_rate-8     1.11s ± 3%     0.85s ± 8%  -23.15%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
NativeHistograms/sum-8          415MB ± 0%     414MB ± 0%     ~     (p=0.548 n=5+5)
NativeHistograms/sum_rate-8    1.16GB ± 0%    0.64GB ± 0%  -44.94%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
NativeHistograms/sum-8          5.16M ± 0%     5.16M ± 0%     ~     (p=0.841 n=5+5)
NativeHistograms/sum_rate-8     21.1M ± 0%     10.3M ± 0%  -51.26%  (p=0.008 n=5+5)
```